+++ index.html
@@ -5,12 +5,14 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite + Lit</title>
-    <link rel="stylesheet" href="./src/index.css" />
-    <script type="module" src="/src/my-element.js"></script>
+    <link rel="stylesheet" href="src/index.css" />
+    <link rel="stylesheet" href="./bootstrap/dist/css/bootstrap.min.css">
+    <script type="module" src="main.js" defer></script>
   </head>
+
   <body>
-    <my-element>
-      <h1>Vite + Lit</h1>
-    </my-element>
+    <main class="container">
+      <home-page></home-page>
+    </main>
   </body>
 </html>
