+++ main.js
@@ -1,0 +1,1 @@
+import { Home } from "./src/components/home";
